+++ plugins/standalone/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @hot-updater/standalone
 
+## 0.25.14
+
+### Patch Changes
+
+- 0a49774: fix(standalone): fetch channels from dedicated endpoint instead of paginated bundles
+  - @hot-updater/core@0.25.14
+  - @hot-updater/plugin-core@0.25.14
+
 ## 0.25.13
 
 ### Patch Changes